fix: update tests for Help overlay redesign
Updated all tests to work with the new Help overlay design:
- Removed tests that referenced Mode::Help
- Added new tests for help overlay behavior
- Updated navigation and state management tests
- Fixed all compilation errors related to the removed Help mode
