+++ src/interactive_ratatui/integration_tests.rs
@@ -442,7 +442,7 @@ mod tests {
                 detail_scroll_offset: 0,
                 selected_result: None,
                 truncation_enabled: true,
-                mode_before_help: None,
+                show_help: false,
             },
         };
 
@@ -482,7 +482,7 @@ mod tests {
                 detail_scroll_offset: 0,
                 selected_result: None,
                 truncation_enabled: true,
-                mode_before_help: None,
+                show_help: false,
             },
         };
         history.push(result_detail_state);
@@ -1040,12 +1040,7 @@ mod tests {
     /// Test Ctrl+C works in all modes
     #[test]
     fn test_ctrl_c_in_all_modes() {
-        let modes = vec![
-            Mode::Search,
-            Mode::MessageDetail,
-            Mode::SessionViewer,
-            Mode::Help,
-        ];
+        let modes = vec![Mode::Search, Mode::MessageDetail, Mode::SessionViewer];
 
         for mode in modes {
             let mut app = InteractiveSearch::new(SearchOptions::default());
