+++ src/interactive_ratatui/mod.rs
@@ -23,9 +23,7 @@ pub mod domain;
 pub mod ui;
 
 #[cfg(test)]
-mod help_mode_rendering_test;
-#[cfg(test)]
-mod help_navigation_test;
+mod help_overlay_test;
 #[cfg(test)]
 mod integration_tests;
 // #[cfg(test)]
